% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firedam.R
\docType{data}
\name{firedam}
\alias{firedam}
\title{Distance from a fire versus the damage in thousands of dollars.}
\format{
A data frame with 5 rows and 2 variables:
\describe{
  \item{DISTANCE}{distance from the fire, in miles}
  \item{DAMAGE}{damage from the fire, in thousands of dollars}
  ...
}
}
\source{
Textbook
}
\usage{
firedam
}
\description{
15 lines of data concerning fires and their damage.
}
\keyword{datasets}
