% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{Plot and .csv files for DDT data}
\usage{
myddt(df, SPECIES)
}
\arguments{
\item{df}{the used data frame}

\item{SPECIES}{the specified species}
}
\value{
A plot of DDT data, a .csv file, and list outputs
}
\description{
Plot and .csv files for DDT data
}
\examples{
\dontrun{myddt(df = ddt, species = "CCATFISH")}

}
